+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.4.2 / 5.59.2] - 2022-10-??
 
+### Added
+- added tooltips to resource access modes [#2300](https://github.com/sandboxie-plus/Sandboxie/issues/2300)
+
 ### Fixed
 - fixed wrong button captions in the breakout process page
 - fixed issue with saving box recovery options